Remove init logger from main file to logger config.
